requirements.txt
requests==2.31.0
python-dotenv==1.0.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.114.0
icalendar==5.0.11
pytz==2023.3
